core: pathfinding: refactor away PathfindingEndpoint
We used to have two endpoints for pathfinding, which shared a tiny amount of
logic inside this class. As this does not make sense anymore, merge this code
with the route pathfinding endpoint and move the request data classes into
separate files.
