+++ core/src/main/java/fr/sncf/osrd/api/pathfinding/request/PathfindingWaypoint.java
@@ -1,0 +1,13 @@
+package fr.sncf.osrd.api.pathfinding.request;
+
+import com.squareup.moshi.Json;
+import fr.sncf.osrd.railjson.schema.common.graph.EdgeDirection;
+
+/** Creates a pathfinding waypoint */
+public record PathfindingWaypoint(
+        @Json(name = "track_section")
+        String trackSection,
+        double offset,
+        EdgeDirection direction
+) {
+}
