+++ core/src/main/java/fr/sncf/osrd/api/pathfinding/response/PathfindingResult.java
@@ -1,35 +1,43 @@
-package fr.sncf.osrd.api;
+package fr.sncf.osrd.api.pathfinding.response;
 
 import static fr.sncf.osrd.envelope_sim.TrainPhysicsIntegrator.POSITION_EPSILON;
 
 import com.google.common.collect.HashMultimap;
 import com.squareup.moshi.Json;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.Moshi;
 import fr.sncf.osrd.infra.api.signaling.SignalingInfra;
 import fr.sncf.osrd.infra.api.signaling.SignalingRoute;
-import fr.sncf.osrd.infra.api.tracks.undirected.OperationalPoint;
-import fr.sncf.osrd.infra.api.tracks.undirected.TrackLocation;
 import fr.sncf.osrd.infra.api.tracks.undirected.TrackSection;
 import fr.sncf.osrd.infra.implementation.RJSObjectParsing;
 import fr.sncf.osrd.infra.implementation.tracks.directed.TrackRangeView;
+import fr.sncf.osrd.railjson.schema.common.ID;
 import fr.sncf.osrd.railjson.schema.common.RJSObjectRef;
-import fr.sncf.osrd.railjson.schema.infra.RJSRoute;
-import fr.sncf.osrd.railjson.schema.infra.RJSTrackSection;
 import fr.sncf.osrd.reporting.warnings.Warning;
 import fr.sncf.osrd.reporting.warnings.WarningRecorderImpl;
 import fr.sncf.osrd.utils.geom.LineString;
+import fr.sncf.osrd.utils.geom.Point;
 import fr.sncf.osrd.utils.graph.Pathfinding;
 import java.util.*;
 
 public class PathfindingResult {
+    public static final JsonAdapter<PathfindingResult> adapterResult = new Moshi
+            .Builder()
+            .add(ID.Adapter.FACTORY)
+            .add(new LineString.Adapter())
+            .add(new Point.Adapter())
+            .build()
+            .adapter(PathfindingResult.class)
+            .failOnUnknown();
     @Json(name = "route_paths")
     public final List<RoutePathResult> routePaths = new ArrayList<>();
     @Json(name = "path_waypoints")
     public final List<PathWaypointResult> pathWaypoints = new ArrayList<>();
-    public List<Warning> warnings;
 
     public LineString geographic = null;
     public LineString schematic = null;
+
+    public List<Warning> warnings;
 
     /**
      * pathfindingResultMake is used to create a class PathfindingResult with a good format to be sent to the middleware
@@ -165,7 +173,7 @@ public class PathfindingResult {
                 continue;
 
             // Add the track ranges to the result
-            routeResult.trackSections.add(new PathfindingEndpoint.DirectionalTrackRangeResult(
+            routeResult.trackSections.add(new DirTrackRange(
                     trackSection.getID(),
                     truncatedRange.getStart(),
                     truncatedRange.getStop()
@@ -179,7 +187,7 @@ public class PathfindingResult {
         var geoList = new ArrayList<LineString>();
         var schList = new ArrayList<LineString>();
 
-        PathfindingEndpoint.DirectionalTrackRangeResult previousTrack = null;
+        DirTrackRange previousTrack = null;
         double previousBegin = 0;
         double previousEnd = 0;
 
@@ -262,79 +270,6 @@ public class PathfindingResult {
             return;
         }
         pathWaypoints.add(newStep);
-    }
-
-    /** A single route on the path */
-    @SuppressFBWarnings({"URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD"})
-    public static class RoutePathResult {
-        public final RJSObjectRef<RJSRoute> route;
-        @Json(name = "track_sections")
-        public final List<PathfindingEndpoint.DirectionalTrackRangeResult> trackSections = new ArrayList<>();
-        @Json(name = "signaling_type")
-        public final String signalingType;
-
-        public RoutePathResult(RJSObjectRef<RJSRoute> route, String signalingType) {
-            this.route = route;
-            this.signalingType = signalingType;
-        }
-    }
-
-    /** One waypoint in the path, represents an operational point */
-    public static class PathWaypointResult {
-        /** ID if the operational point */
-        public String id;
-        /** A point is a suggestion if it's not part of the input path and just an OP on the path */
-        public boolean suggestion;
-        /** Track the point is on */
-        public RJSObjectRef<RJSTrackSection> track;
-        /** Offset of the point on the track */
-        public double position;
-        /** Used to sort the waypoint, not a part of the actual response (transient) */
-        private final transient double trackRangeOffset;
-
-        /** Constructor */
-        private PathWaypointResult(
-                String trackID,
-                double offset,
-                boolean suggestion,
-                String opID,
-                double trackRangeOffset
-        ) {
-            this.suggestion = suggestion;
-            this.track = new RJSObjectRef<>(trackID, "TrackSection");
-            this.position = offset;
-            this.id = opID;
-            this.trackRangeOffset = trackRangeOffset;
-        }
-
-        /** Creates a suggested waypoint from an OP */
-        public static PathWaypointResult suggestion(
-                OperationalPoint op, TrackSection trackSection, double trackRangeOffset
-        ) {
-            return new PathWaypointResult(trackSection.getID(), op.offset(), true, op.id(), trackRangeOffset);
-        }
-
-        /** Creates a user defined waypoint */
-        public static PathWaypointResult userDefined(TrackLocation location, double trackRangeOffset) {
-            return new PathWaypointResult(location.track().getID(), location.offset(), false, null, trackRangeOffset);
-        }
-
-
-        /** Check if two steps result are at the same location */
-        public boolean isDuplicate(PathWaypointResult other) {
-            if (!track.equals(other.track))
-                return false;
-            return Math.abs(position - other.position) < 0.001;
-        }
-
-        /** Merge a suggested with a give step */
-        public void merge(PathWaypointResult other) {
-            suggestion &= other.suggestion;
-            if (!other.suggestion)
-                return;
-            position = other.position;
-            id = other.id;
-        }
     }
 }
